+++ tests/Command/ApiToken/ListTest.php
@@ -7,7 +7,7 @@
 
 declare(strict_types = 1);
 
-namespace Nexcess\Sdk\Cli\Command\ApiToken\Tests;
+namespace Nexcess\Sdk\Cli\Tests\Command\ApiToken;
 
 use Nexcess\Sdk\Cli\ {
   Command\ApiToken\ShowList,
